Performs a gene drop simulation based on the provided pedigree information

Performs a gene drop simulation based on the provided pedigree information

geneDrop(ids, sires, dams, gen, genotype = NULL, n = 5000,
  updateProgress = NULL)

Arguments

ids
character vector of IDs for a set of animals.

sires
character vector with IDS of the sires for the set of animals. NA is used for missing sires.

dams
character vector with IDS of the dams for the set of animals. NA is used for missing dams.

gen
integer vector indicating the generation number for each animal.

genotype
is a dataframe containing known genotypes. It has three columns: id, first, and second. The second and third columns contain the integers indicating the observed genotypes.

Currently there is no means of handling knowing only one haplotype. It will be easy to add another column to handle situations where only one allele is observed and it is not known to be homozygous or heterozygous. The new fourth column could have a frequency for homozygosity that could be used in the gene dropping algorithm.

The genotypes are using indirection (integer instead of character) to indicate the genes because the minipulation of character strings was found to take 20-35 times longer to perform.

Adding additional columns to genotype does not significantly affect the time require. Thus, it is convenient to add the corresponding haplotype names to the dataframe using first_name and second_name.

n
integer indicating the number of iterations to simulate. Default is 5000.

updateProgress
function or NULL. If this function is defined, it will be called during each iteration to update a shiny::Progress object.

Value

data.frame id, parent, V1 ... Vn A data.frame providing the maternal and paternal alleles for an animal for each iteration. The first two columns provide the animal's ID and whether the allele came from the sire or dam. These are followed by n columns indicating the allele for that iteration.