Add animals to an existing breeding or form groups group:

groupAddAssign finds the largest group that can be formed by adding unrelated animals from a set of candidate IDs to an existing group, to a new group it has formed from a set of candidate IDs or if more than 1 group is desired, it finds the set of groups with the largest average size. The function implements a maximal independent set (MIS) algorithm to find groups of unrelated animals. A set of animals may have many different MISs of varying sizes, and finding the largest would require traversing all possible combinations of animals. Since this could be very time consuming, this algorithm produces a random sample of the possible MISs, and selects from these. The size of the random sample is determined by the specified number of iterations.

groupAddAssign(candidates, currentGroup = NULL, kmat, ped,
  threshold = 0.015625, ignore = list(c("F", "F")), minAge = 1,
  iter = 1000, numGp = 1, updateProgress = NULL, withKin = FALSE)

Arguments

candidates
character vector of IDs of the animals available for use in the group.

currentGroup
character vector of IDs of animals currently assigned to the group. Defaults to NULL assuming no groups are existant.

kmat
numeric matrix of pairwise kinship values. Rows and columns are named with animal IDs.

ped
dataframe that is the `Pedigree`. It contains pedigree information including the IDs listed in candidates.

threshold
numeric value indicating the minimum kinship level to be considered in group formation. Pairwise kinship below this level will be ignored.

ignore
list of character vectors representing the sex combinations to be ignored. If provided, the vectors in the list specify if pairwise kinship should be ignored between certain sexes. Default is to ignore all pairwise kinship between females.

minAge
integer value indicating the minimum age to consider in group formation. Pairwise kinships involving an animal of this age or younger will be ignored. Default is 1 year.

iter
integer indicating the number of times to perform the random group formation process. Default value is 1000 iterations.

numGp
integer value indicating the number of groups that should be formed from the list of IDs. Default is 1.

updateProgress
function or NULL. If this function is defined, it will be called during each iteration to update a shiny::Progress object.

withKin
logical variable when set to TRUE the kinship matrix for the group is returned along with the group and score. Defaults to not return the kinship matrix. This maintains compatability with earlier versions.

Value

A list with list items group, score and optionally groupKin. The list item group contains a list of the best group(s) produced during the simulation. The list item score provides the score associated with the group(s). The list item groupKin contains the subset of the kinship matrix that is specific for each group formed.

Details

Part of Group Formation