Add animals to an existing breeding group:

Function to find the largest group that can be formed by adding unrelated animals from a set of candidate IDs to an existing group.

groupAddition(candidates, current.group, kmat, ped, threshold = 0.015625,
  ignore = list(c("F", "F")), min.age = 1, iter = 1000,
  updateProgress = NULL)

Arguments

candidates
character vector of IDs of the animals available for use in the group.

current.group
character vector of IDs of animals currently assigned to the group.

kmat
: matrix row and column names: animal IDs Matrix of pairwise kinship values. Rows and columns are named with animal IDs.

ped
Dataframe of pedigree information including the IDs listed in candidates.

threshold
numeric value indicating the minimum kinship level to be considered in group formation. Pairwise kinship below this level will be ignored.

ignore
list of character vectors representing the sex combinations to be ignored. If provided, the vectors in the list specify if pairwise kinship should be ignored between certain sexes. Default is to ignore all pairwise kinship between females.

min.age
integer value indicating the minimum age to consider in group formation. Pairwise kinships involving an animal of this age or younger will be ignored. Default is 1 year.

iter
integer value indicating number of times to perform the random group formation process. Default is 1000 iterations.

updateProgress
optional function (may be NULL) that is called during each iteration to update a shiny::Progress object.

Value

a list with group, which contains a list of the best group(s) produced during the simulation and score, which provides the score associated with the group(s).

Details

The function implements a maximal independent set (MIS) algorithm to find groups of unrelated animals. A set of animals may have many different MISs of varying sizes, and finding the largest would require traversing all possible combinations of animals. Since this could be very time consuming, this algorithm produces a random sample of the possible MISs, and selects from these. The size of the random sample is determined by the specified number of iterations.