Genetic Value Analysis Functions

Primary Data Structure --- Pedigree

Contains studbook information for a number of individuals. ASSUME: All IDs listed in the sire or dam columns must have a row entry in the id column

A Pedigree is a data frame within the R environment with the following possible columns:

id -- character vector with unique identifier for an individual

sire -- character vector with unique identifier for an individual's father (NA if unknown).

dam -- character vector with unique identifier for an individual's mother (NA if unknown).

sex -- factor levels: "M", "F", "U" Sex specifier for an individual

gen -- integer vector with the generation number of the individual

birth -- Date or NA (optional) with the individual's birth date

exit -- Date or NA (optional) with the individual's exit date (death, or departure if applicable)

ancestry -- character vector or NA (optional) that indicates the geographic population to which the individual belongs.

age -- numeric or NA (optional) indicating the individual's current age or age at exit.

population -- logical (optional) Is the id part of the extant population?

origin -- character vector or NA (optional) that indicates the name of the facility that the individual was imported from if other than local. NA indicates the invidual was not imported.

Genetic Value Analysis Functions

meanKinship --- Calculates the mean kinship for each animal in a kinship matrix

calcA --- Calculates a, the number of an individual's alleles that are rare in each simulation.

alleleFreq --- Calculates the count of each allele in the provided vector.

calcFE --- Calculates founder equivalents.

calcFG --- Calculates founder genome equivalents.

calcFEFG--- Returns founder equivalents FE and FG as elements in a list.

calcGU --- Calculates genome uniqueness for each ID that is part of the population.

geneDrop --- Performs a gene drop simulation based on the pedigree information.

chooseAlleles --- Combines two vectors of alleles by randomly selecting one allele or the other at each position.

calcRetention --- Calculates allelic retention.

filterKinMatrix --- Filters a kinship matrix to include only the egos listed in 'ids'

kinship --- Generates a kinship matrix

reportGV --- Generates a genetic value report for a provided pedigree.

See also